Main: Move activity classes to view package to organize
