+++ app/src/main/AndroidManifest.xml
@@ -12,17 +12,17 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme.NoActionBar"
         android:usesCleartextTraffic="true">
-        <activity android:name=".SplashActivity">
+        <activity android:name=".view.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity" />
-        <activity android:name=".RegisterActivity"/>
-        <activity android:name=".ClassListActivity"/>
-        <activity android:name=".LoginActivity" />
+        <activity android:name=".view.MainActivity" />
+        <activity android:name=".view.RegisterActivity"/>
+        <activity android:name=".view.ClassListActivity"/>
+        <activity android:name=".view.LoginActivity" />
     </application>
 
 </manifest>
