+++ app/src/main/java/com/kangdroid/k_labapplication/view/InfoFragment.kt
@@ -1,4 +1,4 @@
-package com.kangdroid.k_labapplication
+package com.kangdroid.k_labapplication.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,11 +6,19 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.kangdroid.k_labapplication.R
 import com.kangdroid.k_labapplication.databinding.FragmentInfoBinding
 
 class InfoFragment(val num:Int) : Fragment() {
     var binding: FragmentInfoBinding? = null
-    val list = arrayListOf<Int>(R.drawable.info0,R.drawable.info1, R.drawable.info2, R.drawable.info3, R.drawable.info4, R.drawable.info5)
+    val list = arrayListOf<Int>(
+        R.drawable.info0,
+        R.drawable.info1,
+        R.drawable.info2,
+        R.drawable.info3,
+        R.drawable.info4,
+        R.drawable.info5
+    )
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
