+++ app/src/main/java/com/kangdroid/k_labapplication/view/MainActivity.kt
@@ -1,11 +1,7 @@
-package com.kangdroid.k_labapplication
+package com.kangdroid.k_labapplication.view
 
-import android.graphics.PorterDuff
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.core.content.ContextCompat
-import com.google.android.material.tabs.TabLayout
-import com.google.android.material.tabs.TabLayoutMediator
 import com.kangdroid.k_labapplication.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
